+++ src/components/Button/Button.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import { UUI } from '../../utils/uui';
+import { omit } from 'lodash';
 
 
 export interface BaseButtonProps extends React.ButtonHTMLAttributes<HTMLButtonElement> {
@@ -14,7 +15,7 @@ export const Button = UUI.FunctionComponent({
   const { Root } = nodes
 
   return (
-    <Root {...props} className={"u-border u-border-black u-py-1 u-px-2"}>
+    <Root {...omit(props, 'customize', 'className', 'style')} className={"u-border u-border-black u-py-1 u-px-2"}>
       {props.children}
     </Root>
   )